Figure Controllers

+++ src/main/java/celinederoland/repositories/Repository.java
@@ -4,7 +4,7 @@ import java.util.List;
 
 public interface Repository<T> {
 
-    public T findOne(int id);
+    public T findOne(Long id);
 
     public List<T> findAll();
